Update to support laravel 5.4+
Change `share` method to `singleton` because `share` is deleted in Laravel 5.4+ versions.

In [Laravel 5.3 to 5.4 upgrade guide](https://laravel.com/docs/5.4/upgrade#upgrade-5.4.0) we can find this text:

### `share` Method Removed

The share method has been removed from the container. This was a legacy method that has not been documented in several years. If you are using this method, you should begin using the singleton method instead:

```php
$container->singleton('foo', function () {
    return 'foo';
});
```